+++ varnish-macros/src/gen_func.rs
@@ -206,7 +206,7 @@ impl FuncProcessor {
         (format!("  {td_name} *{};\n", self.names.f_fn_name()), decl)
     }
 
-    #[allow(clippy::too_many_lines)]
+    #[expect(clippy::too_many_lines)]
     fn do_fn_param(&mut self, func_info: &FuncInfo, arg_info: &ParamTypeInfo) {
         let arg_name_ident = arg_info.ident.to_ident();
         let arg_value = Self::get_arg_value(func_info, &arg_name_ident);